Move component to design system folder + add storybook stories
