+++ web/src/components/layouts/spinner.tsx
@@ -1,4 +1,4 @@
-import { LangfuseIcon } from "@/src/components/LangfuseLogo";
+import { LangfuseIcon } from "@/src/components/design-system/LangfuseIcon/LangfuseIcon";
 
 export function Spinner(props: { message: string }) {
   return (
