+++ web/src/features/evals/components/maintainer-tooltip.tsx
@@ -1,4 +1,4 @@
-import { LangfuseIcon } from "@/src/components/LangfuseLogo";
+import { LangfuseIcon } from "@/src/components/design-system/LangfuseIcon/LangfuseIcon";
 import {
   Tooltip,
   TooltipTrigger,
